fastapi-intro/requirements.txt
anyio==3.6.1
autopep8==1.7.0
bcrypt==4.0.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.85.0
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.9.1
pydantic==1.10.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
toml==0.10.2
typing_extensions==4.4.0
uvicorn==0.18.3
